Briefly cleaned up files and reorganized folder and file structures for our test server and portara tool
