+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "portara",
-  "version": "1.0.0",
+  "version": "1.0.5",
   "description": "Portara",
   "main": "index.js",
   "scripts": {
